datastorage: refactor to support scalability testing
